@jreback 2021-08-13
does this allow us to get rid of some cython code (that i wrote a long time ago) to do unstacking?