@jbrockmendel 2021-08-18
Yes, just fewer times

On Wed, Aug 18, 2021 at 8:20 AM Jeff Reback ***@***.***>
wrote:

> ***@***.**** requested changes on this pull request.
> ------------------------------
>
> In pandas/core/reshape/reshape.py
> <https://github.com/pandas-dev/pandas/pull/43025#discussion_r691343014>:
>
> > @@ -287,7 +287,7 @@ def get_new_values(self, values, fill_value=None):
>
>          return new_values, new_mask
>
> -    def get_new_columns(self, value_columns):
> +    def get_new_columns(self, value_columns: Index | None):
>
> is this still called?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/pandas-dev/pandas/pull/43025#pullrequestreview-733035402>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AB5UM6E6WA7525MXYRE3WV3T5PFVRANCNFSM5CEFXWSQ>
> .
> Triage notifications on the go with GitHub Mobile for iOS
> <https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675>
> or Android
> <https://play.google.com/store/apps/details?id=com.github.android&utm_campaign=notification-email>
> .
>